Fix discord import conflict by renaming package
